Refactor: move commands, queries and responses to abstractions
