+++ Source/Application/Kysect.Shreks.Application.Abstractions/Github/Queries/GetAssignmentByBranchAndSubjectCourse.cs
@@ -1,0 +1,10 @@
+using MediatR;
+
+namespace Kysect.Shreks.Application.Abstractions.Github.Queries;
+
+public static class GetAssignmentByBranchAndSubjectCourse
+{
+    public record Query(string BranchName, Guid SubjectCourseId) : IRequest<Response>;
+
+    public record Response(Guid AssignmentId);
+}
