+++ Source/Application/Kysect.Shreks.Application.Abstractions/Github/Queries/GetSubjectCourseByOrganization.cs
@@ -1,0 +1,10 @@
+using MediatR;
+
+namespace Kysect.Shreks.Application.Abstractions.Github.Queries;
+
+public static class GetSubjectCourseByOrganization
+{
+    public record Query(string OrganizationName) : IRequest<Response>;
+
+    public record Response(Guid SubjectCourseId);
+}
