+++ Source/Application/Kysect.Shreks.Application.Abstractions/Github/Queries/GetUserByUsername.cs
@@ -1,0 +1,10 @@
+using MediatR;
+
+namespace Kysect.Shreks.Application.Abstractions.Github.Queries;
+
+public static class GetUserByUsername
+{
+    public record Query(string Username) : IRequest<Response>;
+
+    public record Response(Guid UserId);
+}
