+++ Source/Application/Kysect.Shreks.Application.Abstractions/Submissions/Commands/CreateSubmission.cs
@@ -1,0 +1,10 @@
+using MediatR;
+
+namespace Kysect.Shreks.Application.Abstractions.Submissions.Commands;
+
+public static class CreateSubmissionCommand
+{
+    public record Command(Guid StudentId, Guid AssignmentId, string SubmissionUrl) : IRequest<Response>;
+
+    public record Response(Guid SubmissionId);
+}
