+++ Source/Application/Kysect.Shreks.Application.Abstractions/Submissions/Commands/UpdateSubmissionDate.cs
@@ -1,0 +1,8 @@
+using MediatR;
+
+namespace Kysect.Shreks.Application.Abstractions.Submissions.Commands;
+
+public static class UpdateSubmissionDate
+{
+    public record Command(Guid SubmissionId, DateTime NewDate) : IRequest;
+}
