+++ Source/Application/Kysect.Shreks.Application.Abstractions/Submissions/Commands/UpdateSubmissionPoints.cs
@@ -1,0 +1,8 @@
+using MediatR;
+
+namespace Kysect.Shreks.Application.Abstractions.Submissions.Commands;
+
+public static class UpdateSubmissionPoints
+{
+    public record Command(Guid SubmissionId, double NewPoints) : IRequest;
+}
