+++ Source/Application/Kysect.Shreks.Application.Abstractions/Submissions/Commands/UpdateSubmissionUrl.cs
@@ -1,0 +1,8 @@
+using MediatR;
+
+namespace Kysect.Shreks.Application.Abstractions.Submissions.Commands;
+
+public static class UpdateSubmissionUrl
+{
+    public record Command(Guid SubmissionId, string NewUrl) : IRequest;
+}
